Changed register offsets.

+++ musashi/m68k_in.c
@@ -1218,7 +1218,7 @@ M68KMAKE_OP(adda, 16, ., a)
 M68KMAKE_OP(adda, 16, ., .)
 {
 	uint* r_dst = &AX;
-        uint src    = MAKE_INT_16(M68KMAKE_GET_OPER_AY_16);
+        short src   = MAKE_INT_16(M68KMAKE_GET_OPER_AY_16);
 
 	*r_dst = MASK_OUT_ABOVE_32(*r_dst + src);
 }
@@ -1243,7 +1243,7 @@ M68KMAKE_OP(adda, 32, ., a)
 M68KMAKE_OP(adda, 32, ., .)
 {
 	uint* r_dst = &AX;
-        uint src    = M68KMAKE_GET_OPER_AY_32;
+        int src     = M68KMAKE_GET_OPER_AY_32;
 
 	*r_dst = MASK_OUT_ABOVE_32(*r_dst + src);
 }
@@ -9441,7 +9441,7 @@ M68KMAKE_OP(suba, 16, ., a)
 M68KMAKE_OP(suba, 16, ., .)
 {
 	uint* r_dst = &AX;
-	uint src    = MAKE_INT_16(M68KMAKE_GET_OPER_AY_16);
+	short src   = MAKE_INT_16(M68KMAKE_GET_OPER_AY_16);
 
 	*r_dst = MASK_OUT_ABOVE_32(*r_dst - src);
 }
@@ -9466,7 +9466,7 @@ M68KMAKE_OP(suba, 32, ., a)
 M68KMAKE_OP(suba, 32, ., .)
 {
 	uint* r_dst = &AX;
-	uint src    = M68KMAKE_GET_OPER_AY_32;
+	int src     = M68KMAKE_GET_OPER_AY_32;
 
 	*r_dst = MASK_OUT_ABOVE_32(*r_dst - src);
 }
